improving zfp getitem machiney

+++ plugins/codecs/zfp/zfp-private.h
@@ -120,7 +120,7 @@ static int32_t deserialize_meta(uint8_t *smeta, uint32_t smeta_len, int8_t *ndim
     return 0;
 }
 
-static void index_unidim_to_multidim(int8_t ndim, int64_t *shape, int64_t i, int64_t *index) {
+static void index_unidim_to_multidim(uint8_t ndim, int32_t *shape, int64_t i, int64_t *index) {
     int64_t strides[ZFP_MAX_DIM];
     strides[ndim - 1] = 1;
     for (int j = ndim - 2; j >= 0; --j) {
